added group cycle timer to mcu functions

+++ src/node.h
@@ -59,7 +59,7 @@ struct Node {
     double tmp_start_time;
     int dest_node;
     int broadcaster;
-    int group_cycle_start;
+    unsigned long group_cycle_start;
 };
 
 int initialize_nodes(struct Node*);
